+++ src/Engine/Base/Camera.cpp
@@ -48,6 +48,10 @@ Camera::Camera(glm::vec3 position, glm::vec3 up, float yaw, float pitch)
             Position -= Right * velocity;
         if (direction == RIGHT)
             Position += Right * velocity;
+        if (direction == UP)
+            Position += WorldUp * velocity;
+        if (direction == DOWN)
+            Position -= WorldUp * velocity;
     }
 
     // processes input received from a mouse input system. Expects the offset value in both the x and y direction.